Add files via upload
This is an r markdown, which is my capstone project. The "business task" is that Pixar Studios wants to increase its box office revenues for its future production. Looking at the Pixar Studios movie dataset, analyze some trends that could lead to an increase in box office revenue. The data is from Sandip Devre on Kaggle on an Open Database License. This business task is made up for the purpose of practicing a case study.